Remove old jaxb classes, refactor to v3 pkg, rename dir to src-generated
