--- stroom-stats-schema/src/main/java/stroom/stats/schema/ObjectFactory.java
@@ -1,122 +1,0 @@
-/*
- * Copyright 2017 Crown Copyright
- *
- * This file is part of Stroom-Stats.
- *
- * Stroom-Stats is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Stroom-Stats is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Stroom-Stats.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2016.12.22 at 12:28:41 PM GMT 
-//
-
-
-package stroom.stats.schema;
-
-import javax.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the statistics._3 package. 
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: statistics._3
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link Statistics }
-     * 
-     */
-    public Statistics createStatistics() {
-        return new Statistics();
-    }
-
-    /**
-     * Create an instance of {@link CompoundIdentifierType }
-     * 
-     */
-    public CompoundIdentifierType createCompoundIdentifierType() {
-        return new CompoundIdentifierType();
-    }
-
-    /**
-     * Create an instance of {@link Statistics.Statistic }
-     * 
-     */
-    public Statistics.Statistic createStatisticsStatistic() {
-        return new Statistics.Statistic();
-    }
-
-    /**
-     * Create an instance of {@link TagType }
-     * 
-     */
-    public TagType createTagType() {
-        return new TagType();
-    }
-
-    /**
-     * Create an instance of {@link CompoundIdentifierType.LongIdentifier }
-     * 
-     */
-    public CompoundIdentifierType.LongIdentifier createCompoundIdentifierTypeLongIdentifier() {
-        return new CompoundIdentifierType.LongIdentifier();
-    }
-
-    /**
-     * Create an instance of {@link CompoundIdentifierType.StringIdentifier }
-     * 
-     */
-    public CompoundIdentifierType.StringIdentifier createCompoundIdentifierTypeStringIdentifier() {
-        return new CompoundIdentifierType.StringIdentifier();
-    }
-
-    /**
-     * Create an instance of {@link Statistics.Statistic.Tags }
-     * 
-     */
-    public Statistics.Statistic.Tags createStatisticsStatisticTags() {
-        return new Statistics.Statistic.Tags();
-    }
-
-    /**
-     * Create an instance of {@link Statistics.Statistic.Identifiers }
-     * 
-     */
-    public Statistics.Statistic.Identifiers createStatisticsStatisticIdentifiers() {
-        return new Statistics.Statistic.Identifiers();
-    }
-
-}
