--- stroom-stats-schema/src/main/java/stroom/stats/schema/Statistics.java
@@ -1,483 +1,0 @@
-/*
- * Copyright 2017 Crown Copyright
- *
- * This file is part of Stroom-Stats.
- *
- * Stroom-Stats is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Stroom-Stats is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Stroom-Stats.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2016.12.22 at 12:28:41 PM GMT 
-//
-
-
-package stroom.stats.schema;
-
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlSchemaType;
-import javax.xml.bind.annotation.XmlType;
-import javax.xml.datatype.XMLGregorianCalendar;
-
-
-/**
- * <p>Java class for anonymous complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType>
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="statistic" maxOccurs="unbounded">
- *           &lt;complexType>
- *             &lt;complexContent>
- *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *                 &lt;sequence>
- *                   &lt;element name="name" type="{statistics:3}RestrictedStringType"/>
- *                   &lt;element name="time" type="{statistics:3}DateTimeSimpleType"/>
- *                   &lt;choice>
- *                     &lt;element name="count">
- *                       &lt;simpleType>
- *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}long">
- *                           &lt;minInclusive value="1"/>
- *                         &lt;/restriction>
- *                       &lt;/simpleType>
- *                     &lt;/element>
- *                     &lt;element name="value" type="{http://www.w3.org/2001/XMLSchema}double"/>
- *                   &lt;/choice>
- *                   &lt;element name="tags" minOccurs="0">
- *                     &lt;complexType>
- *                       &lt;complexContent>
- *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *                           &lt;sequence>
- *                             &lt;element name="tag" type="{statistics:3}TagType" maxOccurs="unbounded"/>
- *                           &lt;/sequence>
- *                         &lt;/restriction>
- *                       &lt;/complexContent>
- *                     &lt;/complexType>
- *                   &lt;/element>
- *                   &lt;element name="identifiers" minOccurs="0">
- *                     &lt;complexType>
- *                       &lt;complexContent>
- *                         &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *                           &lt;sequence>
- *                             &lt;element name="compoundIdentifier" type="{statistics:3}CompoundIdentifierType" maxOccurs="unbounded"/>
- *                           &lt;/sequence>
- *                         &lt;/restriction>
- *                       &lt;/complexContent>
- *                     &lt;/complexType>
- *                   &lt;/element>
- *                 &lt;/sequence>
- *               &lt;/restriction>
- *             &lt;/complexContent>
- *           &lt;/complexType>
- *         &lt;/element>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = {
-    "statistic"
-})
-@XmlRootElement(name = "statistics")
-public class Statistics implements Serializable {
-
-    @XmlElement(required = true)
-    protected List<Statistics.Statistic> statistic;
-
-    /**
-     * Gets the value of the statistic property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a <CODE>set</CODE> method for the statistic property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getStatistic().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link Statistics.Statistic }
-     * 
-     * 
-     */
-    public List<Statistics.Statistic> getStatistic() {
-        if (statistic == null) {
-            statistic = new ArrayList<>();
-        }
-        return this.statistic;
-    }
-
-
-    /**
-     * <p>Java class for anonymous complex type.
-     * 
-     * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
-     * <pre>
-     * &lt;complexType>
-     *   &lt;complexContent>
-     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
-     *       &lt;sequence>
-     *         &lt;element name="name" type="{statistics:3}RestrictedStringType"/>
-     *         &lt;element name="time" type="{statistics:3}DateTimeSimpleType"/>
-     *         &lt;choice>
-     *           &lt;element name="count">
-     *             &lt;simpleType>
-     *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}long">
-     *                 &lt;minInclusive value="1"/>
-     *               &lt;/restriction>
-     *             &lt;/simpleType>
-     *           &lt;/element>
-     *           &lt;element name="value" type="{http://www.w3.org/2001/XMLSchema}double"/>
-     *         &lt;/choice>
-     *         &lt;element name="tags" minOccurs="0">
-     *           &lt;complexType>
-     *             &lt;complexContent>
-     *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
-     *                 &lt;sequence>
-     *                   &lt;element name="tag" type="{statistics:3}TagType" maxOccurs="unbounded"/>
-     *                 &lt;/sequence>
-     *               &lt;/restriction>
-     *             &lt;/complexContent>
-     *           &lt;/complexType>
-     *         &lt;/element>
-     *         &lt;element name="identifiers" minOccurs="0">
-     *           &lt;complexType>
-     *             &lt;complexContent>
-     *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
-     *                 &lt;sequence>
-     *                   &lt;element name="compoundIdentifier" type="{statistics:3}CompoundIdentifierType" maxOccurs="unbounded"/>
-     *                 &lt;/sequence>
-     *               &lt;/restriction>
-     *             &lt;/complexContent>
-     *           &lt;/complexType>
-     *         &lt;/element>
-     *       &lt;/sequence>
-     *     &lt;/restriction>
-     *   &lt;/complexContent>
-     * &lt;/complexType>
-     * </pre>
-     * 
-     * 
-     */
-    @XmlAccessorType(XmlAccessType.FIELD)
-    @XmlType(name = "", propOrder = {
-        "name",
-        "time",
-        "count",
-        "value",
-        "tags",
-        "identifiers"
-    })
-    public static class Statistic {
-
-        @XmlElement(required = true)
-        protected String name;
-        @XmlElement(required = true)
-        @XmlSchemaType(name = "dateTime")
-        protected XMLGregorianCalendar time;
-        protected Long count;
-        protected Double value;
-        protected Statistics.Statistic.Tags tags;
-        protected Statistics.Statistic.Identifiers identifiers;
-
-        /**
-         * Gets the value of the name property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link String }
-         *     
-         */
-        public String getName() {
-            return name;
-        }
-
-        /**
-         * Sets the value of the name property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link String }
-         *     
-         */
-        public void setName(String value) {
-            this.name = value;
-        }
-
-        /**
-         * Gets the value of the time property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link XMLGregorianCalendar }
-         *     
-         */
-        public XMLGregorianCalendar getTime() {
-            return time;
-        }
-
-        /**
-         * Sets the value of the time property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link XMLGregorianCalendar }
-         *     
-         */
-        public void setTime(XMLGregorianCalendar value) {
-            this.time = value;
-        }
-
-        /**
-         * Gets the value of the count property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link Long }
-         *     
-         */
-        public Long getCount() {
-            return count;
-        }
-
-        /**
-         * Sets the value of the count property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link Long }
-         *     
-         */
-        public void setCount(Long value) {
-            this.count = value;
-        }
-
-        /**
-         * Gets the value of the value property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link Double }
-         *     
-         */
-        public Double getValue() {
-            return value;
-        }
-
-        /**
-         * Sets the value of the value property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link Double }
-         *     
-         */
-        public void setValue(Double value) {
-            this.value = value;
-        }
-
-        /**
-         * Gets the value of the tags property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link Statistics.Statistic.Tags }
-         *     
-         */
-        public Statistics.Statistic.Tags getTags() {
-            return tags;
-        }
-
-        /**
-         * Sets the value of the tags property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link Statistics.Statistic.Tags }
-         *     
-         */
-        public void setTags(Statistics.Statistic.Tags value) {
-            this.tags = value;
-        }
-
-        /**
-         * Gets the value of the identifiers property.
-         * 
-         * @return
-         *     possible object is
-         *     {@link Statistics.Statistic.Identifiers }
-         *     
-         */
-        public Statistics.Statistic.Identifiers getIdentifiers() {
-            return identifiers;
-        }
-
-        /**
-         * Sets the value of the identifiers property.
-         * 
-         * @param value
-         *     allowed object is
-         *     {@link Statistics.Statistic.Identifiers }
-         *     
-         */
-        public void setIdentifiers(Statistics.Statistic.Identifiers value) {
-            this.identifiers = value;
-        }
-
-
-        /**
-         * <p>Java class for anonymous complex type.
-         * 
-         * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
-         * <pre>
-         * &lt;complexType>
-         *   &lt;complexContent>
-         *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
-         *       &lt;sequence>
-         *         &lt;element name="compoundIdentifier" type="{statistics:3}CompoundIdentifierType" maxOccurs="unbounded"/>
-         *       &lt;/sequence>
-         *     &lt;/restriction>
-         *   &lt;/complexContent>
-         * &lt;/complexType>
-         * </pre>
-         * 
-         * 
-         */
-        @XmlAccessorType(XmlAccessType.FIELD)
-        @XmlType(name = "", propOrder = {
-            "compoundIdentifier"
-        })
-        public static class Identifiers {
-
-            @XmlElement(required = true)
-            protected List<CompoundIdentifierType> compoundIdentifier;
-
-            /**
-             * Gets the value of the compoundIdentifier property.
-             * 
-             * <p>
-             * This accessor method returns a reference to the live list,
-             * not a snapshot. Therefore any modification you make to the
-             * returned list will be present inside the JAXB object.
-             * This is why there is not a <CODE>set</CODE> method for the compoundIdentifier property.
-             * 
-             * <p>
-             * For example, to add a new item, do as follows:
-             * <pre>
-             *    getCompoundIdentifier().add(newItem);
-             * </pre>
-             * 
-             * 
-             * <p>
-             * Objects of the following type(s) are allowed in the list
-             * {@link CompoundIdentifierType }
-             * 
-             * 
-             */
-            public List<CompoundIdentifierType> getCompoundIdentifier() {
-                if (compoundIdentifier == null) {
-                    compoundIdentifier = new ArrayList<>();
-                }
-                return this.compoundIdentifier;
-            }
-
-        }
-
-
-        /**
-         * <p>Java class for anonymous complex type.
-         * 
-         * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
-         * <pre>
-         * &lt;complexType>
-         *   &lt;complexContent>
-         *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
-         *       &lt;sequence>
-         *         &lt;element name="tag" type="{statistics:3}TagType" maxOccurs="unbounded"/>
-         *       &lt;/sequence>
-         *     &lt;/restriction>
-         *   &lt;/complexContent>
-         * &lt;/complexType>
-         * </pre>
-         * 
-         * 
-         */
-        @XmlAccessorType(XmlAccessType.FIELD)
-        @XmlType(name = "", propOrder = {
-            "tag"
-        })
-        public static class Tags {
-
-            @XmlElement(required = true)
-            protected List<TagType> tag;
-
-            /**
-             * Gets the value of the tag property.
-             * 
-             * <p>
-             * This accessor method returns a reference to the live list,
-             * not a snapshot. Therefore any modification you make to the
-             * returned list will be present inside the JAXB object.
-             * This is why there is not a <CODE>set</CODE> method for the tag property.
-             * 
-             * <p>
-             * For example, to add a new item, do as follows:
-             * <pre>
-             *    getTag().add(newItem);
-             * </pre>
-             * 
-             * 
-             * <p>
-             * Objects of the following type(s) are allowed in the list
-             * {@link TagType }
-             * 
-             * 
-             */
-            public List<TagType> getTag() {
-                if (tag == null) {
-                    tag = new ArrayList<>();
-                }
-                return this.tag;
-            }
-
-        }
-
-    }
-
-}
