+++ stroom-stats-service/src/integration-test/java/stroom/stats/correlation/QueryResource_simpleQueries_IT.java
@@ -20,8 +20,8 @@ import stroom.query.api.v2.TableSettingsBuilder;
 import stroom.stats.AbstractAppIT;
 import stroom.stats.api.StatisticType;
 import stroom.stats.configuration.StatisticConfiguration;
-import stroom.stats.schema.ObjectFactory;
-import stroom.stats.schema.Statistics;
+import stroom.stats.schema.v3.ObjectFactory;
+import stroom.stats.schema.v3.Statistics;
 import stroom.stats.shared.EventStoreTimeIntervalEnum;
 import stroom.stats.test.QueryApiHelper;
 import stroom.stats.test.StatisticsHelper;
