+++ stroom-stats-service/src/main/java/stroom/stats/streams/StatisticsFlatMappingStreamFactory.java
@@ -32,8 +32,8 @@ import stroom.stats.hbase.EventStoreTimeIntervalHelper;
 import stroom.stats.hbase.HBaseStatisticConstants;
 import stroom.stats.partitions.StatKeyPartitioner;
 import stroom.stats.properties.StroomPropertyService;
-import stroom.stats.schema.ObjectFactory;
-import stroom.stats.schema.Statistics;
+import stroom.stats.schema.v3.ObjectFactory;
+import stroom.stats.schema.v3.Statistics;
 import stroom.stats.shared.EventStoreTimeIntervalEnum;
 import stroom.stats.streams.aggregation.StatAggregate;
 import stroom.stats.streams.mapping.AbstractStatisticFlatMapper;
