+++ stroom-stats-service/src/main/java/stroom/stats/streams/mapping/AbstractStatisticFlatMapper.java
@@ -29,9 +29,9 @@ import stroom.stats.hbase.HBaseStatisticConstants;
 import stroom.stats.hbase.uid.UID;
 import stroom.stats.hbase.uid.UniqueIdCache;
 import stroom.stats.properties.StroomPropertyService;
-import stroom.stats.schema.CompoundIdentifierType;
-import stroom.stats.schema.Statistics;
-import stroom.stats.schema.TagType;
+import stroom.stats.schema.v3.CompoundIdentifierType;
+import stroom.stats.schema.v3.Statistics;
+import stroom.stats.schema.v3.TagType;
 import stroom.stats.shared.EventStoreTimeIntervalEnum;
 import stroom.stats.streams.StatKey;
 import stroom.stats.streams.StatisticWrapper;
