+++ stroom-stats-service/src/test/java/stroom/stats/streams/mapping/TestValueStatToAggregateFlatMapper.java
@@ -37,8 +37,8 @@ import stroom.stats.hbase.uid.MockUniqueIdCache;
 import stroom.stats.hbase.uid.UID;
 import stroom.stats.hbase.uid.UniqueIdCache;
 import stroom.stats.properties.MockStroomPropertyService;
-import stroom.stats.schema.Statistics;
-import stroom.stats.schema.TagType;
+import stroom.stats.schema.v3.Statistics;
+import stroom.stats.schema.v3.TagType;
 import stroom.stats.shared.EventStoreTimeIntervalEnum;
 import stroom.stats.streams.StatKey;
 import stroom.stats.streams.StatisticWrapper;
